backend: move HAR capture buffer/entry logic into backend/harcapture
Per review: the buffer/entry-building/redaction logic (harcapture.go and its
tests) has no dependency on the backend package, so it moves into its own
backend/harcapture package instead of adding ~700 lines to backend directly.
Buffer, NewBuffer, AddEntry, Len, ToHARString, and DrainRequestBody are now
its exported surface.

The thin HandlerMiddleware/Handler glue (har_capture_middleware.go) stays in
package backend: it implements backend interfaces and is wired directly into
defaultHandlerMiddlewares in serve.go, which backend/harcapture can't import
back without a cycle.

Also responding to review re: req.GetHTTPHeader for the capture header check
-- kept the existing raw map read. Grafana core sets it with a plain,
unprefixed map assignment (the same producer contract FromAlertHeaderName
relies on), and GetHTTPHeader only surfaces Authorization/X-Id-Token/Cookie
or "http_"-prefixed keys, so switching would make the header lookup always
miss and silently disable capture.
