+++ backend/harcapture/redaction_test.go
@@ -1,4 +1,4 @@
-package backend
+package harcapture
 
 import (
 	"net/http"
